@@ -1,7 +1,7 @@
 {
 	"name": "misskey",
 	"author": "syuilo <i@syuilo.com>",
-	"version": "11.25.1-20190716154445",
+	"version": "11.26.2",
 	"codename": "daybreak",
 	"repository": {
 		"type": "git",
@@ -37,7 +37,6 @@
 		"@fortawesome/free-solid-svg-icons": "5.9.0",
 		"@fortawesome/vue-fontawesome": "0.1.6",
 		"@koa/cors": "3.0.0",
-		"@typescript-eslint/parser": "1.11.0",
 		"@types/bcryptjs": "2.4.2",
 		"@types/bull": "3.5.15",
 		"@types/cbor": "2.0.0",
@@ -82,7 +81,7 @@
 		"@types/ratelimiter": "2.1.28",
 		"@types/redis": "2.8.13",
 		"@types/rename": "1.0.1",
-		"@types/request": "2.48.1",
+		"@types/request": "2.48.2",
 		"@types/request-promise-native": "1.0.16",
 		"@types/request-stats": "3.0.0",
 		"@types/rimraf": "2.0.2",
@@ -99,8 +98,9 @@
 		"@types/webpack-stream": "3.2.10",
 		"@types/websocket": "0.0.40",
 		"@types/ws": "6.0.1",
+		"@typescript-eslint/parser": "1.11.0",
 		"animejs": "3.0.1",
-		"apexcharts": "3.8.2",
+		"apexcharts": "3.8.3",
 		"autobind-decorator": "2.4.0",
 		"autosize": "4.0.2",
 		"autwh": "0.1.0",
@@ -116,18 +116,19 @@
 		"commander": "2.20.0",
 		"content-disposition": "0.5.3",
 		"crc-32": "1.2.0",
-		"css-loader": "3.0.0",
+		"css-loader": "3.1.0",
 		"cssnano": "4.1.10",
 		"dateformat": "3.0.3",
 		"deep-equal": "1.0.1",
 		"diskusage": "1.1.3",
 		"double-ended-queue": "2.1.0-0",
 		"emojilib": "2.4.0",
-		"eslint": "6.0.1",
+		"eslint": "6.1.0",
 		"eslint-plugin-vue": "5.2.3",
 		"eventemitter3": "4.0.0",
 		"feed": "3.0.0",
 		"file-type": "12.0.1",
+		"fluent-ffmpeg": "2.1.2",
 		"gulp": "4.0.2",
 		"gulp-cssnano": "2.1.3",
 		"gulp-imagemin": "6.0.0",
@@ -157,7 +158,7 @@
 		"koa-compress": "3.0.0",
 		"koa-favicon": "2.0.1",
 		"koa-json-body": "5.3.0",
-		"koa-logger": "3.2.0",
+		"koa-logger": "3.2.1",
 		"koa-mount": "4.0.0",
 		"koa-multer": "1.0.2",
 		"koa-router": "7.4.0",
@@ -169,13 +170,12 @@
 		"lolex": "4.1.0",
 		"lookup-dns-cache": "2.1.0",
 		"minio": "7.0.10",
-		"mocha": "6.1.4",
+		"mocha": "6.2.0",
 		"moji": "0.5.1",
-		"moment": "2.24.0",
 		"ms": "2.1.2",
 		"nested-property": "1.0.1",
 		"node-fetch": "2.6.0",
-		"nodemailer": "6.2.1",
+		"nodemailer": "6.3.0",
 		"nprogress": "0.2.0",
 		"object-assign-deep": "0.4.0",
 		"os-utils": "0.0.14",
@@ -192,7 +192,7 @@
 		"pug": "2.0.4",
 		"punycode": "2.1.1",
 		"pureimage": "0.1.6",
-		"qrcode": "1.4.0",
+		"qrcode": "1.4.1",
 		"random-seed": "0.3.0",
 		"randomcolor": "0.5.4",
 		"ratelimiter": "3.3.0",
@@ -218,7 +218,7 @@
 		"stylus": "0.54.5",
 		"stylus-loader": "3.0.2",
 		"summaly": "2.3.0",
-		"systeminformation": "4.14.3",
+		"systeminformation": "4.14.4",
 		"syuilo-password-strength": "0.0.1",
 		"terser-webpack-plugin": "1.3.0",
 		"textarea-caret": "3.1.0",
@@ -229,14 +229,13 @@
 		"tslint": "5.18.0",
 		"tslint-sonarts": "1.9.0",
 		"typeorm": "0.2.18",
-		"typescript": "3.5.2",
+		"typescript": "3.5.3",
 		"uglify-es": "3.3.9",
 		"ulid": "2.3.0",
-		"url-loader": "2.0.1",
+		"url-loader": "2.1.0",
 		"uuid": "3.3.2",
 		"v-animate-css": "0.0.3",
 		"v-debounce": "0.1.2",
-		"video-thumbnail-generator": "1.1.3",
 		"vue": "2.6.10",
 		"vue-color": "2.7.0",
 		"vue-content-loading": "1.6.0",
@@ -244,7 +243,7 @@
 		"vue-i18n": "8.12.0",
 		"vue-js-modal": "1.3.31",
 		"vue-json-pretty": "1.6.0",
-		"vue-loader": "15.7.0",
+		"vue-loader": "15.7.1",
 		"vue-marquee-text-component": "1.1.1",
 		"vue-prism-component": "1.1.1",
 		"vue-router": "3.0.7",
@@ -257,10 +256,13 @@
 		"vuex": "3.1.1",
 		"vuex-persistedstate": "2.5.4",
 		"web-push": "3.3.5",
-		"webpack": "4.35.3",
-		"webpack-cli": "3.3.5",
+		"webpack": "4.36.1",
+		"webpack-cli": "3.3.6",
 		"websocket": "1.0.29",
-		"ws": "7.1.0",
+		"ws": "7.1.1",
 		"xev": "2.0.1"
+	},
+	"devDependencies": {
+		"@types/fluent-ffmpeg": "2.1.10"
 	}
 }
